fix: update ToolErrorTracker to use usize for counts and limits

+++ crates/forge_domain/src/tool_call.rs
@@ -187,17 +187,18 @@ where
         })
 }
 
-#[derive(Clone, Debug)]
+#[derive(Default, Clone, Debug)]
 pub struct ToolErrorTracker {
-    counts: HashMap<ToolName, u32>,
-    limit: u32,
+    counts: HashMap<ToolName, usize>,
+    limit: usize,
 }
 
 impl ToolErrorTracker {
-    pub fn new(limit: u32) -> Self {
+    pub fn new(limit: usize) -> Self {
         Self { counts: Default::default(), limit }
     }
 
+    // FIXME: add a new function that takes tool call records as input
     pub fn adjust(&mut self, failed: &[&ToolName], succeeded: &[&ToolName]) -> &mut Self {
         // Handle failures first
         let uniq_failed = failed.iter().collect::<HashSet<&&ToolName>>();
@@ -226,15 +227,15 @@ impl ToolErrorTracker {
             .collect::<Vec<_>>()
     }
 
-    pub fn get_counts(&self) -> &HashMap<ToolName, u32> {
+    pub fn get_counts(&self) -> &HashMap<ToolName, usize> {
         &self.counts
     }
 
-    pub fn get_attempt_count(&self, tool_name: &ToolName) -> u32 {
+    pub fn get_attempt_count(&self, tool_name: &ToolName) -> usize {
         *self.counts.get(tool_name).unwrap_or(&0)
     }
 
-    pub fn get_attempts_remaining(&self, tool_name: &ToolName) -> u32 {
+    pub fn get_attempts_remaining(&self, tool_name: &ToolName) -> usize {
         let current_attempts = self.get_attempt_count(tool_name);
         self.limit.saturating_sub(current_attempts)
     }
